+++ includes/JtlConnectorAdmin.php
@@ -1984,6 +1984,7 @@ final class JtlConnectorAdmin //phpcs:ignore PSR1.Classes.ClassDeclaration.Missi
             case '1.39.7':
             case '1.39.8':
             case '1.39.9':
+            case '1.39.10':
             default:
                 self::activate_linking();
         }