+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is the changelog of our "JTL WooCommerce Connector".
 ## Unreleased
 - CO-2263 - fix empty delivery addresses
 - CO-2389 - fix delivery status
+- CO-2409 - support Additional Variation Images Gallery plugin
 
 ## 1.39.9 _2023-06-13_
 - CO-2230 - fix Invisibility Type